add token and user extractor middleware

+++ part4/controllers/blogs.js
@@ -1,3 +1,4 @@
+const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
@@ -6,12 +7,11 @@ blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
     .find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
-})
+}) 
 
 blogsRouter.post('/', async (request, response) => {
-  const { title, author, url, likes, userId } = request.body
-
-  const user = await User.findById(userId)
+  const { title, author, url, likes } = request.body
+  const user = request.user
 
   if (!title || !url) {
     return response.status(400).json({ error: 'title or url missing'})
@@ -33,7 +33,12 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+
+  if ( blog.user.toString() === user._id.toString() ) {
+    await Blog.findByIdAndDelete(request.params.id)
+  }
   response.status(204).end()
 })
